requirements.txt
openai~=1.30.1
python-dotenv~=1.0.1
chromadb~=0.5.0
pdfplumber~=0.11.1
jieba~=0.42.1